requirements.txt
scikit-learn==1.4.2
pandas==2.2.2
fastapi==0.111.0
gunicorn==22.0.0
pymongo==4.7.2
python-dotenv==1.0.1
pydantic==2.7.1